+++ components/Pagination/PaginationItem.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '../../utils/utils';
+import { cn } from '../../utils/css';
 
 const PaginationItem = React.forwardRef<HTMLLIElement, React.ComponentProps<'li'>>(
   ({ className, ...props }, ref) => <li ref={ref} className={cn('', className)} {...props} />
