+++ components/Pagination/PaginationLink.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { cn } from '../../utils/utils';
+import { cn } from '../../utils/css';
 import { ButtonProps, buttonVariants } from './ButtonPagination';
 
 type PaginationLinkProps = {
